Enable unit tests to run standalone
Enable unit tests to be run standalane; that is, without having a
local copy of the horizon source.  This is primarly intended for use in
automated builds.  The approach, taken from tuskar-ui, is to add an
entry in test-requirements.txt that refers to a source tarball of
horizon which is expanded into the virtual environment, permitting
references to horizon and openstack_dashboard to resolve correctly.

Add additional tooling to support running tests and building virtual
environments to mirror what is done in horizon.