Fix improperly rendered Edit Share modal form
Edit Share form validation works for only one time. On the second try
whatever values are used, a correct ones or not, the form closes with
"There was an error submitting the form. Please try again." message.
It appears that form_invalid method returns broken submit url.
This patch moves rendering to get_initial method to get a proper
submit url for valid and invalid forms.

Change-Id: I85a18da3a72cb8d964e0372eb5ede1ae9c94bb59
Closes-Bug: #1585252