Only show available share protocols on share creation modal
Do not show share protocols in Shares drop-down menu if
the backend is not available.
This patch set is a workaround for this bug. We are using the
settings override feature. Operators will need to place a
'enabled_share_protocols: ['proto1', 'proto2', ...]' setting
on the OPENSTACK_MANILA_SETTINGS files.
The correct fix for this bug would be to add a capabilities
endpoint in the controller plane in order to get the available
backends.

Change-Id: I13637491994e6f074e66fe0f3026a303944892ec
Closes-Bug: #1622732