Fix migration-start id-dependent form
Share migration migration-start form was requiring the IDs
to be inputted for the new_share_type and new_share_network
fields. The main problem was finding the ID for the share type.

This fix redesigns the form to include dropdown lists instead of
text fields, allowing the share type and share network to be
selected from a list instead of having to input their IDs.

Closes-bug: #1626780
Change-Id: I60d00d3e3beb994f46db3543b59fed767b1823cb