Adds Manila-UI Devstack plugin
Manila-UI was being deployed if Manila plugin was enabled.
This is not the most advisable thing since counting on an UI
is not neccesary for the majority of the cases when contributing
code to Manila. This patch-set adds the posibility of deploying
the UI on demand by enabling it as a plugin for Devstack as follows:

enable_plugin manila-ui https://github.com/openstack/manila-ui

Change-Id: Id564e4ea6b675aa197ae377aa7a756e1bd0625b6
Closes-Bug: #1570033
Implements BP add-devstack-plugin