Add access key column to rules table
To align with the functionality offered by the CLI, this patch
adds an "Access Key" column to the rules table.

The access key is also added to the share details overview.

For completeness, also the access level and the state are now
listed in the details overview.

Change-Id: I38516e453e300b4c08a7da5de4acf09d2bdac4d2
Closes-Bug: #1638934
(cherry picked from commit 1e6f23cd7cc66ce7222ffef48d79048bfbfbb774)