Ignore current share's size in extend quota bar
The progress bar displaying the currently used quota in the 'Extend Share'
panel includes the size of the share to be extended. While this is not
wrong in itself, the share's new size is then added on top which double-counts
the share's quota and leads to a wrongly displayed progress bar.

This patch proposes to remove the current share's size from the used
quota, which avoids the mentioned issue and also aligns this panel with
what is displayed in other projects when extending the size of a resource.

Change-Id: If82ed6f45c6d1f364624dfbf54f2112fcd8093df
Closes-Bug: #1650292