Use proper horizon version for unit tests
Before, it was installing latest master version for unit tests.
But it is incorrect version of horizon for any of stable branch of our
manila Ui plugin. So, update test-requirements.txt file to install
appropriate branch of Horizon to avoid issues related to evolvement of
latest Horizon that can be incompatible with stable version of plugin.

Change-Id: Iadc26ac3997711143ca27de92a25656c96d5416c
Closes-Bug: #1664213