Use EXTRA_TABS/EXTRA_STEPS mechanism to handle manila quotas
The previous manila-ui approach to show shared-file-system default
quotas is too hacky and fragile as the bug reveals.
This commit applies EXTRA_TABS/EXTRA_STEPS plugin interface
introduced to horizon in Rocky to show the manila quota tab
in the default quotas table and quota related forms.

monkey-patching from manila quotas has been dropped.

Note that EXTRA_STEPS feature is being reviewd in horizon
https://review.openstack.org/#/c/560679/ so workflow steps related to
manila in the project quota form and the default quota form are not
shown until the corresponding horizon feature is merged.
However, this commit will unblock manila-ui CI so I believe this is
a good compromise.

Note that manila support in the project overview panel is dropped
temporarily as further refactoring in the project overview panel
is being done in horizon side (which will land in Rocky-2).

Closes-Bug: #1759340
Co-Authored-By: Victoria Martinez de la Cruz <victoria@redhat.com>
Change-Id: I22ed7d757c5e5e902f1e85a15c33b34de6c609f1