Require the (soon-to-be) Rocky version of Horizon
... or at least the first published tarball which will lead to Rocky.
On the other side this means forcing the usage of mox as long as
there are tests that depends on it (the switch was already flipped
in Horizon, so it needs to be enabled here).

This allows us to depend on newer Horizon features, including changes
to the test suite that are useful for the mox removal effort.

Also remove mox since there are no unit tests that need it.

Change-Id: Ia8e0d5512f11e3d191ca2f6259075e35490290c2
Signed-off-by: Charles Short <zulcss@gmail.com>