Show share network details to non-admin
Horizon attempts to show associated share servers when displaying
share network details in the project share-networks dashboard.  But
unless the horizon user is an admin this generates a 403 exception
which in turn causes Horizon to pop-up an error message about being
unable to retrieve share network details.

So instead catch exceptions collecting share server information in
this context and continue on gracefully, presenting the share server
details that are available.

Change-Id: I822ba95dc8e41df2775f908c287565f5da6a7e9d
Closes-Bug: #1702396