Updated to get quotas data for Modify Quotas dialog Share tab
Current Modfiy Quotas action of a project in Identity->Projects
for manila ui does not get quotas data for the project. If user
updated quotas for manila ui, the saved data will not be shown
in the Modify Quotas dialog Share tab.

This commit includes:
1) Added a call to get quotas data and also maps the data
fields retuned to the UI data fields.
2) Added unit tests.
3) Added a release note.

Change-Id: I5fb9cef577b109530fde9a4bafe930ea05f3fed8
Closes-bug: #1842119
Depends-on: https://review.opendev.org/#/c/679513/