Changes to Share Update Functionality in Manila UI
In the manila_api module, there is a check for is_public
param as if it was a string, but it is always a boolean,
so this control is not necessary and needs to be removed.
Removed control for is_public param in share_update in manila_api.
Updated forms and view to expect Boolean not a string in
manila_dashboards module.
Made changes to tests module to work with Boolean,
not a string.

Closes-Bug: #1896533

Change-Id: Iac94219b883b5fd9a55453cdcf6fa9b6f540e63d