Set parent share az when creating a share from snapshot
This patch-set modifies how we manage user input when creating a share
from snapshot. Since we use the same form for the basic create share workflow
and for the create share from snapshot workflow, we need to add a check
to distinguish both cases: when creating a share from snapshot, the
parent share availability zone is picked by default. Otherwise, if it's a
regular share create, the user can pick up the availability zone they need.

Closes-Bug: #1915706

Change-Id: I8d93c2214cb187a763df861555ddd7b444c9ca89