Drop lower constraints testing
The lower-constraints job is failing because
the lower-constraints.txt file is out of date
with reality - a new pip dependency resolver [1]
now lets us know that several packages have
conflicting lower bounds. There is no gain
expected if we retroactively attempt to fix
lower constraints for required packages in this
older stable branch - deployers have already
wrangled through the requirements to identify
what versions of packages are needed for
manila_ui to work.

[1] http://pyfound.blogspot.com/2020/11/pip-20-3-new-resolver.html
[2] https://pip.pypa.io/en/stable/user_guide/#changes-to-the-pip-dependency-resolver-in-20-3-2020
[3] http://lists.openstack.org/pipermail/openstack-discuss/2021-March/021204.html

Change-Id: I5382eb49901436fa38a0c076060cd4a4033a2182
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>