tox: Declare bash as allowlist_externals
tox4 checks allowlist_externals more strictly and
all externals need to be specified explicitly.

skipsdist=True is also droppped. When skipsdist=True is specified,
a target application is not installed into a tox env, but there
seems no reason to do so. In various other OpenStack projects,
a target applicatin needs to be installed into a tox vnev, for
example, in docs and/or linter jobs and it causes failures with tox4.
I believe dropping it would avoid potential future failures.

Change-Id: I56d2a654f2be87e98026dfa237d5c81d5fcefa37